requirements.txt
python-dotenv==0.21.0
python-telegram-bot==20.0
openai==0.25.0
